refactor: roles components refactored
