+++ src/api/users.api.ts
@@ -7,12 +7,7 @@ const useAllUsers = () => {
     queryKey: [API_KEYS.ALL_USERS],
     queryFn: async () => {
       const usersData = await sdk.users.getUsers();
-
-      return usersData.map((user) => ({
-        ...user,
-        created_at: new Date(user.created_at),
-        updated_at: new Date(user.updated_at),
-      }));
+      return usersData;
     },
     refetchOnWindowFocus: false,
   });
